requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
requests==2.31.0
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
google-api-python-client==2.130.0
networkx==3.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
praw==7.7.1
youtube-transcript-api==0.6.2
plotly==5.22.0
